requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Jinja2==3.1.2
Flask-MySQLdb==1.0.1
MySQLdb==2.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.1
